# openapi/astrocollab.yaml
...
openapi: 3.1.0
info:
  title: AstroCollab API
  version: 0.2.0-draft.1
  license:
    name: MIT
    identifier: MIT
  description: 'Draft API for collaborative astrophotography. A telescope''s software signs in, describes

    the rig, asks what to shoot tonight and reports what it captured. This version adopts the

    protocol Starfront already speaks. See spec/protocol.md for the rules behind each route.'
...
paths:
  /api/v1/agent/hello:
    post:
      operationId: hello
      summary: Say hello and describe the rig
      description: 'Says what this telescope is and, if it likes, where it points. Also a heartbeat. Send
        it

        before asking for work and whenever the rig changes.'
      parameters: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              examples:
                example:
                  value:
                    agent: '000000000001'
                    name: Vega 530
                    protocol: 1
                    serverTime: 1791171023.0
        '401':
          description: Missing, unknown or revoked token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflicts with how things stand, such as a rig that cannot meet a project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: The body does not match its type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
      tags:
      - telescope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelloRequest'
            examples:
              example:
                value:
                  protocol: 1
                  profile:
                    name: Vega 530
                    focalLength: 530.0
                    pixelSize: 3.76
                    sensorWidth: 6248
                    sensorHeight: 4176
                    binning: 1
                    filters:
                      Ha: 7.0
                      OIII: 7.0
                      SII: 7.0
                      L: null
                    colour: false
                    rotation: null
                    typicalHfr: 2.4
                    typicalGuideRms: 0.62
                    exposures:
                      Ha: 300.0
                      OIII: 300.0
                      SII: 300.0
                      L: 120.0
                    hoursPerNight: 6.0
                    windowFrom: '21:30'
                    windowTo: 04:30
                    scale: 1.463313962264151
                    field:
                    - 2.53966267672956
                    - 1.6974441962264148
                  presence:
                    ra: 0.7123
                    dec: 41.27
                    state: imaging
                    target: M31 halo in narrowband
                    project: '000000000002'
  /api/v1/agent/projects:
    get:
      operationId: openProjects
      summary: Browse open projects
      description: Every open project, and whether this telescope can help each one.
      parameters: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenProjects'
              examples:
                example:
                  value:
                    projects:
                    - id: '000000000003'
                      name: M51 in LRGB
                      coordinator: coordinator
                      ownerId: ''
                      region:
                        ra: 202.4696
                        dec: 47.1952
                        width: 0.25
                        height: 0.2
                        rotation: 0.0
                      kind: single
                      requirements:
                        minFocalLength: null
                        maxFocalLength: 400.0
                        minScale: null
                        maxScale: null
                        acceptColour: true
                        colourMaxMoon: null
                        maxHfr: 3.0
                        maxGuideRms: null
                        minExposure: null
                        maxExposure: null
                        filters:
                          L: null
                          R: null
                          G: null
                          B: null
                        maxMoonIllumination: null
                        minMoonSeparation: null
                        minAltitude: null
                        requireCalibrated: false
                        minFramesPerVisit: 10
                      goals:
                        L: 20.0
                        R: 5.0
                        G: 5.0
                        B: 5.0
                      notes: ''
                      compatibility:
                        ok: false
                        certain: false
                        checks:
                        - check: focal length
                          ok: false
                          detail: 530 mm is longer than the 400 mm the project wants
                        - check: scale
                          ok: true
                          detail: 1.46"/px
                        - check: filters
                          ok: false
                          detail: can shoot L — but no R; no G; no B
                        - check: stars
                          ok: true
                          detail: usually 2.40", project wants 3" or better
                        summary: 'cannot contribute: 530 mm is longer than the 400 mm the project wants;
                          can shoot L — but no R; no G; no B'
                      joined: false
                      collected: {}
                      participants: 0
                      participantsOnline: 0
                      participantNames: []
                    - id: '000000000002'
                      name: M31 halo in narrowband
                      coordinator: coordinator
                      ownerId: ''
                      region:
                        ra: 10.6847
                        dec: 41.269
                        width: 7.0
                        height: 4.5
                        rotation: 35.0
                      kind: mosaic
                      requirements:
                        minFocalLength: null
                        maxFocalLength: null
                        minScale: null
                        maxScale: null
                        acceptColour: true
                        colourMaxMoon: null
                        maxHfr: 3.5
                        maxGuideRms: null
                        minExposure: 120.0
                        maxExposure: 600.0
                        filters:
                          H: 7.0
                          O: 7.0
                        maxMoonIllumination: null
                        minMoonSeparation: 30.0
                        minAltitude: 30.0
                        requireCalibrated: false
                        minFramesPerVisit: 10
                      goals:
                        H: 10.0
                        O: 10.0
                      notes: Faint H-alpha and OIII around M31. Calibrated subs only.
                      compatibility:
                        ok: true
                        certain: true
                        checks:
                        - check: scale
                          ok: true
                          detail: 1.46"/px
                        - check: filters
                          ok: true
                          detail: can shoot H, O
                        - check: exposures
                          ok: true
                          detail: the default exposures suit the project
                        - check: stars
                          ok: true
                          detail: usually 2.40", project wants 3.5" or better
                        summary: can contribute
                      joined: false
                      collected: {}
                      participants: 0
                      participantsOnline: 0
                      participantNames: []
                    protocol: 1
        '401':
          description: Missing, unknown or revoked token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: The body does not match its type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
      tags:
      - telescope
  /api/v1/agent/projects/{project_id}/join:
    post:
      operationId: joinProject
      summary: Join a project
      description: 'Takes a share of a project, checked against the profile the rig last sent. The share
        is

        the whole region tiled with this rig''s own camera, at its own sub lengths. Joining twice

        returns the share already held. Joining is consent: the share arrives accepted. A rig that

        has not described its optics gets 400; one that cannot meet the rules gets 409.'
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResponse'
              examples:
                example:
                  value:
                    task:
                      id: '000000000004'
                      project: '000000000002'
                      projectName: M31 halo in narrowband
                      agent: '000000000001'
                      region:
                        ra: 10.6847
                        dec: 41.269
                        width: 7.0
                        height: 4.5
                        rotation: 35.0
                      filters:
                      - filter: H
                        exposure: 300.0
                        hours: 10.0
                      - filter: O
                        exposure: 300.0
                        hours: 10.0
                      state: accepted
                      version: 2
                      issued: 1791171029.0
                      note: joined by Vega 530
                      seconds: 72000.0
                      cells:
                      - ra: 7.648994769812747
                        dec: 39.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 0
                        column: 0
                      - ra: 10.6847
                        dec: 39.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 0
                        column: 1
                      - ra: 13.720405230187254
                        dec: 39.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 0
                        column: 2
                      - ra: 7.580299025898128
                        dec: 41.269
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 1
                        column: 0
                      - ra: 10.6847
                        dec: 41.269
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 1
                        column: 1
                      - ra: 13.789100974101872
                        dec: 41.269
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 1
                        column: 2
                      - ra: 7.506193403991872
                        dec: 42.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 2
                        column: 0
                      - ra: 10.6847
                        dec: 42.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 2
                        column: 1
                      - ra: 13.863206596008128
                        dec: 42.769
                        width: 2.53966267672956
                        height: 1.6974441962264148
                        rotation: 35.0
                        row: 2
                        column: 2
                      share:
                      - 0
                      - 1
                      - 2
                      - 3
                      - 4
                      - 5
                      kind: mosaic
                      visit:
                        seconds: 3300.0
                        frames:
                          H: 11
                        filter: H
                      assignedNight: ''
                      assignedAt: 1791171030.0
                      dealtHours: 6.0
                      tiledRotation: null
                    requirements:
                      minFocalLength: null
                      maxFocalLength: null
                      minScale: null
                      maxScale: null
                      acceptColour: true
                      colourMaxMoon: null
                      maxHfr: 3.5
                      maxGuideRms: null
                      minExposure: 120.0
                      maxExposure: 600.0
                      filters:
                        H: 7.0
                        O: 7.0
                      maxMoonIllumination: null
                      minMoonSeparation: 30.0
                      minAltitude: 30.0
                      requireCalibrated: false
                      minFramesPerVisit: 10
        '400':
          description: The request cannot be served as asked yet, such as joining before describing the
            rig.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, unknown or revoked token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such thing, or not this caller's.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflicts with how things stand, such as a rig that cannot meet a project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                example:
                  value:
                    detail: 'cannot contribute: 1000 mm is longer than the 400 mm the project wants; can
                      shoot L — but no R; no G; no B'
        '422':
          description: The body does not match its type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
      tags:
      - telescope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequest'
            examples:
              example:
                value:
                  hours: 0.0
                  exposure: 0.0
                  exposures:
                    Ha: 300.0
                    OIII: 300.0
                    SII: 300.0
                    L: 120.0
  /api/v1/agent/report:
    post:
      operationId: report
      summary: Report what was shot
      description: 'Hands back what was shot, one record per night, filter and panel, and hears whether
        each

        counts. Reporting the same night, filter and panel again keeps the larger figure.'
      parameters: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              examples:
                example:
                  value:
                    recorded:
                    - id: '000000000005'
                      accepted: true
                      duplicate: false
                      verdict:
                        accepted: true
                        reasons: []
                        unverified: []
                        summary: accepted
                    - id: '000000000006'
                      accepted: true
                      duplicate: false
                      verdict:
                        accepted: true
                        reasons: []
                        unverified: []
                        summary: accepted
        '401':
          description: Missing, unknown or revoked token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: The caller may not do this.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: The body does not match its type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
      tags:
      - telescope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
            examples:
              example:
                value:
                  contributions:
                  - project: '000000000002'
                    task: '000000000004'
                    night: '2026-10-05'
                    panel: '0'
                    filterName: OIII
                    frames: 11
                    seconds: 3300.0
                    exposure: 300.0
                    footprint:
                      ra: 7.648994769812747
                      dec: 39.769
                      width: 2.53966267672956
                      height: 1.6974441962264148
                      rotation: 35.0
                    scale: 1.463313962264151
                    focalLength: 530.0
                    colour: false
                    hfr: 2.3413023396226413
                    guideRms: 0.58
                    bandpass: 7.0
                  - project: '000000000002'
                    task: '000000000004'
                    night: '2026-10-05'
                    panel: '1'
                    filterName: OIII
                    frames: 11
                    seconds: 3300.0
                    exposure: 300.0
                    footprint:
                      ra: 10.6847
                      dec: 39.769
                      width: 2.53966267672956
                      height: 1.6974441962264148
                      rotation: 35.0
                    scale: 1.463313962264151
                    focalLength: 530.0
                    colour: false
                    hfr: 2.4876337358490566
                    guideRms: 0.58
                    bandpass: 7.0
...
  /api/v1/auth:
    get:
      operationId: authStatus
      summary: Can people sign in here
      parameters: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatus'
              examples:
                example:
                  value:
                    discord: true
                    guild: '000000000000000001'
                    roleRequired: false
                    publicUrl: https://collab.example
      tags:
      - signin
      security:
      - {}
...
  /api/v1/auth/me:
    get:
      operationId: authMe
      summary: Who is signed in
      parameters: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
              examples:
                example:
                  value:
                    id: '000000000000000042'
                    name: Vega Observatory
                    admin: false
                    canStart: true
        '401':
          description: Missing, unknown or revoked token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: The body does not match its type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
      tags:
      - signin
...
security:
- bearerAuth: []
components:
  schemas:
...
    ErrorBody:
      type: object
      required:
      - detail
      properties:
        detail:
          type: string
      description: An error. `detail` says what went wrong, in words to show the operator.
...
    HelloResponse:
      type: object
      required:
      - agent
      - name
      - protocol
      - serverTime
      properties:
        agent:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        protocol:
          type: integer
          format: int32
        serverTime:
          $ref: '#/components/schemas/UnixTime'
    Id:
      type: string
      pattern: ^[0-9a-f]{12}$
      description: 'A server-made identifier: 12 lowercase hexadecimal characters.'
...
    OpenProjects:
      type: object
      required:
      - projects
      - protocol
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/OpenProject'
        protocol:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: Bearer
      description: 'Send `Authorization: Bearer <token>`. Telescope routes take a telescope''s agent token;

        account routes take a person''s token from sign-in. The two are never interchangeable.'
servers:
- url: https://collab.example
  description: Reserved example; use the address of the server you join.
  variables: {}